Manual merging differences between v2.0.1 and v2.0.0 done in hidden r792 to v3.0.0 branch. Fixes #91, fixes #92, fixes #93, fixes #89.
Actually, this should be a task of the developer, who has never started with the version merges. I personally think it's an absolute cheek, let us sit with this promised follow-up work. Sorry, but I wasted so much time now to look for the differences between the both versions in SVN and(!) Git.