EZComments is using full qualified URLs in the database.
That is what I don't really like. I think we should not store the URL to an object. It is better to store the arguments separately, so that everytime we can generate a valid URL. If you change your domain or the displayname of a module, then there are many broken URLs stored in the database. An idea for the future.